分类删除

+++ src/main/java/com/bc/pmpheep/back/service/ProductTypeServiceImpl.java
@@ -52,14 +52,16 @@ public class ProductTypeServiceImpl implements ProductTypeService {
         ResponseBean responseBean = new ResponseBean();
         int typeType = productType.getTypeType();
         int count = 0;
-        /*if (typeType==1){ //学科分类
+        if (typeType==1){ //学科分类
+
             count = productTypeDao.deleteSubjectTypeById(productType.getId());
         }else if(typeType==2){ //内容分类
+
             count = productTypeDao.deleteLeafContentTypeById(productType.getId());
             productTypeDao.refreshLeafOfContentType();
         }else{
             //TODO 如果后续有新增其他分类...
-        }*/
+        }
